Move IdentifierFactory to base test factories
