+++ tests/factories.py
@@ -50,6 +50,7 @@ from website.project.sanctions import (
 )
 from website.notifications.model import NotificationSubscription, NotificationDigest
 from website.archiver.model import ArchiveTarget, ArchiveJob
+from website.identifiers.model import Identifier
 from website.archiver import ARCHIVER_SUCCESS
 from website.project.licenses import NodeLicense, NodeLicenseRecord, ensure_licenses
 ensure_licenses = functools.partial(ensure_licenses, warn=False)
@@ -769,6 +770,15 @@ class NodeLicenseRecordFactory(ModularOdmFactory):
             )
         )
         return super(NodeLicenseRecordFactory, cls)._create(*args, **kwargs)
+
+
+class IdentifierFactory(ModularOdmFactory):
+    class Meta:
+        model = Identifier
+
+    referent = SubFactory(RegistrationFactory)
+    category = 'carpid'
+    value = 'carp:/24601'
 
 
 def render_generations_from_parent(parent, creator, num_generations):
